+++ javascript/experiments/use-webpack-and-swc/src/index.js
@@ -1,5 +1,6 @@
 import { createRoot } from 'react-dom/client';
-import { App } from './App';
+
+import { App } from './components/App';
 
 import './index.scss';
